+++ build.mjs
@@ -18,7 +18,7 @@ const builds = [
   { format: 'esm', outfile: 'dist/acmeticker.esm.js' },
   { format: 'esm', minify: true, outfile: 'dist/acmeticker.esm.min.js' },
   { format: 'cjs', outfile: 'dist/acmeticker.cjs' },
-  { format: 'cjs', minify: true, outfile: 'dist/acmeticker.cjs.min.js' },
+  { format: 'cjs', minify: true, outfile: 'dist/acmeticker.min.cjs' },
   {
     format: 'iife',
     globalName: 'AcmeTicker',